Fix sash position... now in Windows.
This is again a great example why wxWidgets sucks: if the splitter position is
set greater than the size of the frame, the SetSashSize() call will be silently
ignored.

This would not be such a problem if this behavior worked the same way in all
platforms (it would be sooo easier to debug). However the frame creation
procedure is very different in all platforms:

- on OSX when the frame is loaded from XRC, it has size immediately
  (or at least, SetSashPosition just works)
- on Linux when the frame is loaded from XRC, it is only allocated, but not
  "created". The creation event will be fired before the frame shows up,
  and in the event handler the frame already has size.
- on Windows the creation event is either called immediately after loading
  or is never called: I could not create an event handler on that. However the
  frame has no size after the constructor, so SetSashPosition() doesn't work
  anyway.

Because there is no "show up" event in wxPython, the easiest method was to call
SetSashPosition after the frame.Show() call. Thanks to God it seems to work on
all platforms now.

And if anyone explains why there are so many calls in wxWidgets that actually
do NOTHING in some cases (without throwing an exception or something), I will
be very grateful.